Refactoring Home Screen
- added home_app_bar
- added home_bannner
- added left_section_title
- added services_horizontal_list_view
- added top_categories_grid_view
