+++ lib/core/routing/app_router.dart
@@ -4,7 +4,7 @@ import 'package:waie/core/di/dependency_injection.dart';
 import 'package:waie/core/routing/routes.dart';
 import 'package:waie/features/OnBoarding/landing_screen.dart';
 import 'package:waie/features/OnBoarding/welcome_screen.dart';
-import 'package:waie/features/home/home_screen.dart';
+import 'package:waie/features/home/presentation/home_screen.dart';
 import 'package:waie/features/login/logic/cubit/login_cubit.dart';
 import 'package:waie/features/login/presentation/login_screen.dart';
 import 'package:waie/features/otp/otp_screen.dart';
