+++ lib/features/home/presentation/widgets/left_section_title.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class LeftSectionTitle extends StatelessWidget {
+  final String text;
+
+  const LeftSectionTitle({
+    Key? key,
+    required this.text,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.centerLeft,
+      child: Text(
+        text,
+        style: const TextStyle(
+          fontSize: 22,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+    );
+  }
+}
